requirements.txt
# Framework used
Django

# Manage virtual environment for easy dependency management
virtualenvwrapper-win; platform_system == "Windows"
virtualenvwrapper; platform_system != "Windows"

# Additional Django Modules
django-multiselectfield
django-markdownx
django-jet
Markdown